+++ lib/tutorial_coach_mark.dart
@@ -1,9 +1,11 @@
 library tutorial_coach_mark;
 
 import 'dart:async';
+import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:tutorial_coach_mark/src/target/target_focus.dart';
+import 'package:tutorial_coach_mark/src/util.dart';
 import 'package:tutorial_coach_mark/src/widgets/tutorial_coach_mark_widget.dart';
 
 export 'package:tutorial_coach_mark/src/target/target_content.dart';
@@ -33,6 +35,7 @@ class TutorialCoachMark {
   final bool pulseEnable;
   final Widget? skipWidget;
   final bool showSkipInLastTarget;
+  final ImageFilter? imageFilter;
 
   OverlayEntry? _overlayEntry;
 
@@ -56,6 +59,7 @@ class TutorialCoachMark {
     this.pulseEnable = true,
     this.skipWidget,
     this.showSkipInLastTarget = true,
+    this.imageFilter,
   }) : assert(opacityShadow >= 0 && opacityShadow <= 1);
 
   OverlayEntry _buildOverlay({bool rootOverlay = false}) {
@@ -83,19 +87,47 @@ class TutorialCoachMark {
           finish: finish,
           rootOverlay: rootOverlay,
           showSkipInLastTarget: showSkipInLastTarget,
+          imageFilter: imageFilter,
         );
       },
     );
   }
 
   void show({required BuildContext context, bool rootOverlay = false}) {
-    Future.delayed(Duration.zero, () {
-      if (_overlayEntry == null) {
-        _overlayEntry = _buildOverlay(rootOverlay: rootOverlay);
-        // ignore: invalid_null_aware_operator
-        Overlay.of(context, rootOverlay: rootOverlay)?.insert(_overlayEntry!);
-      }
+    OverlayState? overlay = Overlay.of(context, rootOverlay: rootOverlay);
+    overlay.let((it) {
+      showWithOverlayState(overlay: it, rootOverlay: rootOverlay);
     });
+  }
+
+  // `navigatorKey` needs to be the one that you passed to MaterialApp.navigatorKey
+  void showWithNavigatorStateKey({
+    required GlobalKey<NavigatorState> navigatorKey,
+    bool rootOverlay = false,
+  }) {
+    navigatorKey.currentState?.overlay.let((it) {
+      showWithOverlayState(
+        overlay: it,
+        rootOverlay: rootOverlay,
+      );
+    });
+  }
+
+  void showWithOverlayState({
+    required OverlayState overlay,
+    bool rootOverlay = false,
+  }) {
+    postFrame(() => _createAndShow(overlay, rootOverlay: rootOverlay));
+  }
+
+  void _createAndShow(
+    OverlayState overlay, {
+    bool rootOverlay = false,
+  }) {
+    if (_overlayEntry == null) {
+      _overlayEntry = _buildOverlay(rootOverlay: rootOverlay);
+      overlay.insert(_overlayEntry!);
+    }
   }
 
   void finish() {